feat: add login_required router to projectRouter
also edit Project schema to include timestamps.

+++ back/src/routers/projectRouter.js
@@ -1,6 +1,8 @@
 import is from "@sindresorhus/is";
 import { Router } from "express";
 import { projectService } from "../services/projectService";
+import { Project } from "../db/models/Project";
+import { login_required } from "../middlewares/login_required";
 
 const projectRouter = Router();
 
@@ -27,24 +29,22 @@ projectRouter.get("/user/project", async function (req, res, next) {
 
 
 // 주어진 ID의 User에게 신규 프로젝트 post
-projectRouter.post("/user/project", async function (req, res, next) {
+projectRouter.post("/user/project", login_required, async function (req, res, next) {
   try {
     if (is.emptyObject(req.body)) {
       throw new Error(
         "headers의 Content-Type을 application/json으로 설정해주세요"
       );
     }
-    // User 식별해서 이중 보안 (프론트는 된다!) 넣기!!!!!
-    
     // req (request) 에서 데이터 가져오기
-    const userId = req.body.id;
+    const currentUserId = req.currentUserId;
     const title = req.body.title;
     const startDate = req.body.startDate;
     const endDate = req.body.endDate;
     const description = req.body.description;
 
     // Project DB에 신규 프로젝트 추가하기
-    const newProject = await projectService.addProject({ userId, title, startDate, endDate, description })
+    const newProject = await projectService.addProject({ userId: currentUserId, title, startDate, endDate, description })
     res.status(201).json(newProject);
   } catch (error) {
     next(error);
@@ -53,17 +53,25 @@ projectRouter.post("/user/project", async function (req, res, next) {
 
 
 // 주어진 projectId에 해당되는 프로젝트 수정
-projectRouter.put("/user/project", async function (req, res, next) {
+projectRouter.put("/user/project/:projectId", login_required, async function (req, res, next) {
   try {
     if (is.emptyObject(req.body)) {
       throw new Error(
         "headers의 Content-Type을 application/json으로 설정해주세요"
       );
     }
-    // User 식별해서 이중 보안 (프론트는 된다!) 넣기!!!!!
+
+    // User authentication
+    const currentUserId = req.currentUserId; // 현재 로그인 중인 userId
+    const userId = req.body.userId; // Project 내에 저장된 userId
+    if (currentUserId!==userId) {
+      throw new Error(
+        "currentUser가 본 Project의 작성자가 아닙니다."
+      );
+    }
 
     // req (request) 에서 데이터 가져오기
-    const projectId = req.body.projectId;
+    const projectId = req.params.projectId;
     const title = req.body.title || undefined;
     const startDate = req.body.startDate || undefined;
     const endDate = req.body.endDate || undefined;
@@ -79,18 +87,26 @@ projectRouter.put("/user/project", async function (req, res, next) {
 });
 
 
-// 주어진 project ID에 해당되는 프로젝트 삭제
-projectRouter.delete("/user/project", async function (req, res, next) {
+// 주어진 projectId에 해당되는 프로젝트 삭제
+projectRouter.delete("/user/project/:projectId", login_required, async function (req, res, next) {
   try {
     if (is.emptyObject(req.body)) {
       throw new Error(
         "headers의 Content-Type을 application/json으로 설정해주세요"
       );
     }
-    // User 식별해서 이중 보안 (프론트는 된다!) 넣기!!!!!
+
+    // User authentication
+    const currentUserId = req.currentUserId; // 현재 로그인 중인 userId
+    const userId = req.body.userId; // Project 내에 저장된 userId
+    if (currentUserId!==userId) {
+      throw new Error(
+        "currentUser가 본 Project의 작성자가 아닙니다."
+      );
+    }
 
     // req (request) 에서 데이터 가져오기
-    const projectId = req.body.projectId;
+    const projectId = req.params.projectId;
 
     // Project DB에서 projectId에 매칭되는 Project 삭제하기
     const deleteSuccessful = await projectService.deleteProject({projectId})
